+++ firecode/references.py
@@ -9,10 +9,23 @@ references = {
     "CREST": "Pracht, P.; Bohle, F.; Grimme, S. PCCP 2020, 22, 7169-7192 - https://doi.org/10.1039/C9CP06869D",
     "ALPB": "Ehlert, S.; Stahn, M.; Spicher, S.; Grimme, S. J. Chem. Theory Comput. 2021, 17, 4250-4261 - https://doi.org/10.1021/acs.jctc.1c00471",
     "AIMNET2": "Anstine, D.M.; Zubatyuk, R.; Isayev, O. Chem. Sci., 2025, 16, 10228-10244 - https://doi.org/10.1039/D4SC08572H",
-    "AIMNET2-NSE": "Kalita, B.; Zubatyuk, R.; Anstine D. M.; Bergeler, M.; Settels, V.; Stork, C.; Spicher, S.; Isayev, O ChemRxiv, 2025 - https://doi.org/10.26434/chemrxiv-2025-kdg6n",
+    "AIMNET2-NSE": (
+        "Kalita, B.; Zubatyuk, R.; Anstine D. M.; Bergeler, M.;\n"
+        + "Settels, V.; Stork, C.; Spicher, S.; Isayev, O ChemRxiv, 2025 - https://doi.org/10.26434/chemrxiv-2025-kdg6n"
+    ),
     "UMA": (
-        "Wood, B. M.; Dzamba, M.; Fu, X.; Gao, M.; Shuaibi, M.; Barroso-Luque, L.; Abdelmaqsoud, K.; "
-        + "Gharakhanyan, V.; Kitchin, J. R. ; Levine, D. S.; Michel, K.; Sriram,A. ; Cohen,T. ; Das, A.; "
+        "Wood, B. M.; Dzamba, M.; Fu, X.; Gao, M.; Shuaibi, M.; Barroso-Luque, L.; Abdelmaqsoud, K.;\n"
+        + "Gharakhanyan, V.; Kitchin, J. R. ; Levine, D. S.; Michel, K.; Sriram,A. ; Cohen,T. ; Das, A.;\n"
         + "Rizvi, A. ; Sahoo, S. J.; Ulissi, Z. W.; Zitnick, C. L. ArXiv, 2025 - https://arxiv.org/abs/2506.23971"
+    ),
+    "NEB": (
+        "Henkelman, G.; Jonsson, H. Chem. Phys 2000, 113, 9978 - https://ddx.doi.org/10.1063/1.1323224\n"
+        + "Henkelman, G.; Uberuaga, B. P.; Jonsson, H. Chem. Phys 2000, 113, 9901 - https://ddx.doi.org/10.1063/1.1329672\n"
+        + "Kolsbjerg, E. L.; Groves, M. N.; Hammer, B. J. Chem. Phys 2016, 145, 094107 - https://ddx.doi.org/10.1063/1.4961868\n"
+        + "Makri, S.; Ortner, C.; Kermode, J. R.; J. Chem. Phys. 2019, 150, 094109 - https://dx.doi.org/10.1063/1.5064465"
+    ),
+    "FSM": (
+        "Marks, J.; Gomes, J. J. Chem. Theory Comput. 2025, 21, 12110-12120 - https://doi.org/10.1021/acs.jctc.5c01492\n"
+        + "Marks, J.; Gomes, J. arXiv 2025 - http://arxiv.org/abs/2501.06159"
     ),
 }